Corrected coercion error resulting in clearing field values.
Added conditions to allow for simpleXrm.setAttVal(a, false) and
simpleXrm.setAttVal(a, 0). Previously these conditions would result in
the value of the field being cleared due to JavaScript coercing these
values to be "falsey". This would prevent a developer from properly
setting a boolean field value to false or setting a numeric or money
field to 0 or $0.00.